refactor config macro defs and add test command
Instead of hardcoding every macro definition and the condition for
applying it, now we use the field type and default value to categorize
the macro into one of three options (opt-in, opt-out, value) at
comptime. Additionally, the field names are normalized into what Janet
expects in janetconf.h (such as all-caps, "NO_" prefix for opt-out
config) to avoid having to hardcode the normalized macro names. After
all this, an inline for loop iterates over the fields, and checks its
category (opt-in, opt-out, value) and value (null, true, false) to
determine whether or not to apply the macro.

Also, `zig build test` now works, and runs the Janet test suite using
the freshly compiled interpreter (it uses the same underlying steps
and cached files as `zig build` for this).